requirements.txt
pydantic>=2.6.1
fastapi>=0.115.6
uvicorn>=0.34.0
pydantic-settings>=2.7.1
sqlmodel>=0.0.22
httpx>=0.28.1
pydantic-ai[logfire,google]>=0.2.14
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
sqlalchemy[asyncio]>=2.0.37
pgvector>=0.3.6
voyageai>=0.3.2
tenacity>=9.0.0
alembic>=1.14.1
logfire[fastapi,httpx,sqlalchemy,system-metrics]>=3.12.0
python-docx>=0.8.11
PyPDF2>=3.0.1